Make instances returned by the accessors on USBDevice `===` comparable.
Add new private HeapVector members in USBConfiguration, USBInterface,
and USBAlternateInterface to hold an array of related instances rather
than making a new instance each time the method is invoked, and the
array is constructed in the constructor. By doing this way to make
related instances within USBDevice `===` comparable.

On the WPT webusb side:
- Add new USBDevice SameObject test.
- Add new garbage collection test to improve code coverage.

Bug: 1274922
Change-Id: I96c928c4fef4692d8b74dbaa26b22a6c6ae60443