Support small/large/dynamic/logical viewport units
This CL adds support for new units:

 * svw, svh (etc): Resolves against the viewport as it would be with
                   browser controls fully expanded.
 * lvw, lvh (etc): Resolves against the viewport as it would be with
                   browser controls fully minimized.
 * dvw, dvh (etc): Resolves against the viewport as it presently
                   appears given the current state of browser controls.
 * vi, vb:         Resolves against vw/vh depending on the writing-mode
                   of the target element.

Note that dv* units are deliberately not responsive to browser control
animations, and will always resolve to either sv* or lv* (not some
value in between). This is to avoid excessive scroll-driven relayouts
on mobile devices.

This CL uses the existing system/bits for invalidation of viewport
units. Targeted invalidation of dynamic units will come in a
subsequent CL.

Bug: 1093055
Change-Id: I3ba55a86de53514f2020a2d5bdda579f06dd9d0a