URLPattern: Set unmatched optional groups to undefined instead of ''.
This addresses the issues raised in:

https://github.com/WICG/urlpattern/issues/162

The main changes in this CL are:

1. The webidl is modified to allow passed back undefined.  I'm told
   real webidl should support `(USVString or undefined)` here, but our
   webidl compiler does not support that.  So we use `any` instead.
2. We must examine if the WTF::String is null or not before populating
   the returned array.  If its null, then we convert to undefined
   instead.

It turned out, however, that StringUTF8Adaptor was also converting
empty strings to null strings.  It was necessary to fix this in order
to return undefined and empty string in the correct places in the
returned values.

Bug: 1292699
Change-Id: I27a0be2567bb9ce4592ca15a5216fd5a58bdbf17