[IndexedDB Storage Partition] (1) Tentative WPT
This is a very basic WPT test of indexeddb storage partitioning. It
doesn't work yet in chrome or firefox, but does in safari. We should get
in a test now so that when we can flip the flag for partitioned storage
in chrome, we have a test ready for the virtual suite.

This CL is a part of a series:
(1) Tentative WPT
(2) Add to third-party-storage-partitioning virtual suite

Bug: 1290613
Change-Id: I2dc6c59494aabcbf5c8e084b0b687362f76431a9