AppHistory bfcache support
Currently, with appHistory and bfcache both enabled, appHistory's
entries() arrays are not updated when a navigation is served from
bfcache. It keeps the same appHistory.entries() that it had when the
document was put into bfcache, but the entries() almost certainly have
changed in the meantime.

This CL ensures that when blink is notified of the restore from
bfcache, it also receives updated appHistory.entries().

Fixed: 1241655
Change-Id: I9c49312ccbbec714f751f3b59c8d1f05edb024fd